requirements.txt
# requirements.txt
pytesseract
Pillow
opencv-python
transformers
torch
presidio-analyzer
presidio-anonymizer
presidio-image-redactor
spacy
